// src/Models/api.ts
export interface FormSchemaDTO {
    /**
     * 
     * @type {number}
     * @memberof FormSchemaDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof FormSchemaDTO
     */
    'jsonSchema'?: string;
    /**
     * 
     * @type {string}
     * @memberof FormSchemaDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof FormSchemaDTO
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof FormSchemaDTO
     */
    'lastUpdate'?: string;
}